[Refactor] Spinner 파일 ts 변환
